fix: allow https redirection from www.ioha2020.sg to www.ioha2021.gov.sg
Since NLB has already pointed www.ioha2020.sg to this redirection
server, we have to account for the case of https://www.ioha2020.sg
(which we ordinarily do not care about, since this is taken care of
by KeyCDN with their zone alias - this is what the www domain usually
points to).

As such, we need to add a server name to the ioha2021.gov.sg server
block on the https redirect config. This works because the ioha 2021
SSL cert includes a DNS name for 'www.ioha2020.sg'