refactor: Move config to conf.d/
This commit brings the organisation of our nginx config files in-line
with existing conventions, where nginx.conf will reference all files
in the `conf.d/` directory.

This also moves the default port 80 listener into its own config.
While not necessary, it will bring us further in-line with the config
that we will use in future once we use a Docker image that incorporates
certbot, so that we can support automatic provision and renewal of
LetsEncrypt certs.

- Dockerfile: copy specific config into `conf.d/`
- nginx.conf: Shard out port 80 listener to its own config,
  include files in `conf.d/`