chore: set ssl session timeout to be 1 hour
Currently, our ssl session timeout is for one day. This has
led to a shortage of sessions, which occasionally leads to the
following error:

2021/02/02 01:01:15 [alert] 492#492: *127421 could not allocate
new session in SSL session shared cache "MozSSL"
while SSL handshaking, client: 128.199.23.243, server: 0.0.0.0:443

Running out of sessions is not actually a big problem (see https://trac.nginx.org/nginx/ticket/621) but these errors regularly trigger alerts notification on our
monitoring channel, which leads to notification fatigue.

However, since:
- we don't expect our individual sessions to actually last for a day
- nginx sets a default of 5 minutes for this parameter (http://nginx.org/en/docs/http/ngx_http_ssl_module.html#ssl_session_timeout)

We figured that it would not be a big problem to reduce the value of
the ssl_session_timeout config. To be conservative, we will be setting it
to one hour (as an upper bound on what a reasonable isomer session would be)
which is still significantly more than nginx's default.

Other references:
- https://github.com/mozilla/server-side-tls/issues/198#issuecomment-327402827